@@ -1,5 +1,6 @@
 {
   "name": "portfolio",
+  "license": "MIT",
   "version": "1.0.0",
   "homepage": "https://Adyl.dev",
   "description": "Portfolio site for Adyl ITT0",
@@ -20,6 +21,7 @@
     "eslint-config-next": "^12.2.3",
     "fs-extra": "^10.1.0",
     "globby": "^13.1.2",
+    "next": "^13.1.6",
     "postcss": "^8.4.14",
     "postcss-flexbugs-fixes": "^5.0.2",
     "postcss-preset-env": "^7.7.2",
@@ -40,9 +42,10 @@
     "fathom-client": "^3.5.0",
     "framer-motion": "^6.5.1",
     "mdx-bundler": "^9.0.1",
-    "next": "^12.2.3",
+    "next": "^13.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "sharp": "^0.31.3",
     "three": "^0.143.0",
     "three-stdlib": "^2.13.0"
   },
