@@ -1,39 +1,37 @@
-module.exports = {
-  reactStrictMode: true,
-  trailingSlash: true,
-  pageExtensions: ['page.js', 'api.js'],
-  webpack(config, { isServer }) {
-    // Run custom scripts
-    if (isServer) {
-      require('./scripts/generate-sitemap');
-      require('./scripts/draco');
-    }
+export const reactStrictMode = true;
+export const trailingSlash = true;
+export const pageExtensions = ['page.js', 'api.js'];
+export function webpack(config, { isServer }) {
+  // Run custom scripts
+  if (isServer) {
+    require('./scripts/generate-sitemap');
+    require('./scripts/draco');
+  }
 
-    // Import `svg` files as React components
-    config.module.rules.push({
-      test: /\.svg$/,
-      resourceQuery: { not: [/url/] },
-      use: [{ loader: '@svgr/webpack', options: { svgo: false } }],
-    });
+  // Import `svg` files as React components
+  config.module.rules.push({
+    test: /\.svg$/,
+    resourceQuery: { not: [/url/] },
+    use: [{ loader: '@svgr/webpack', options: { svgo: false } }],
+  });
 
-    // Import videos, models, hdrs, and fonts
-    config.module.rules.push({
-      test: /\.(mp4|hdr|glb|woff|woff2)$/i,
-      type: 'asset/resource',
-    });
+  // Import videos, models, hdrs, and fonts
+  config.module.rules.push({
+    test: /\.(mp4|hdr|glb|woff|woff2)$/i,
+    type: 'asset/resource',
+  });
 
-    // Force url import with `?url`
-    config.module.rules.push({
-      resourceQuery: /url/,
-      type: 'asset/resource',
-    });
+  // Force url import with `?url`
+  config.module.rules.push({
+    resourceQuery: /url/,
+    type: 'asset/resource',
+  });
 
-    // Import `.glsl` shaders
-    config.module.rules.push({
-      test: /\.glsl$/,
-      type: 'asset/source',
-    });
+  // Import `.glsl` shaders
+  config.module.rules.push({
+    test: /\.glsl$/,
+    type: 'asset/source',
+  });
 
-    return config;
-  },
-};
+  return config;
+}
